# api/docs/openapi-full.yaml
openapi: 3.0.3
info:
  title: GlobalReach API
  description: Enterprise Email Marketing Platform API
  version: 2.0.0
  contact:
    email: support@globalreach.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.globalreach.com/api/v1
    description: Production server
  - url: https://staging.api.globalreach.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: AUTHENTICATION_FAILED
        message:
          type: string
          example: Invalid credentials
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    Template:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        description:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /templates:
    get:
      tags: [Templates]
      summary: Get user templates
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
    post:
      tags: [Templates]
      summary: Create template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                subject:
                  type: string
                body:
                  type: string
                description:
                  type: string
                isDefault:
                  type: boolean
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Template'
  /templates/{id}:
    get:
      tags: [Templates]
      summary: Get template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Template'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Templates]
      summary: Update template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                subject:
                  type: string
                body:
                  type: string
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      tags: [Templates]
      summary: Delete template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Template deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...